Move tabulateAlias tests to separate file
